refactor(ProductController,JdbcProductDAO): add an exception handler for DAO related action
